# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Coupon",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "zoheirmaz.zm@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
        "/coupon/create": {
            "post": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Login a user",
                "requestBody": {
                    "request": "Pet",
                    "description": "create coupon input data structure",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "integer"
                                    },
                                    "rules": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "rule": {
                                                    "type": "integer"
                                                },
                                                "values": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "results": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "result": {
                                                    "type": "integer"
                                                },
                                                "values": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "login sucessful"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/coupon/rules-results": {
            "get": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Get coupons rules and results",
                "responses": {
                    "200": {
                        "description": "requst is sucessful"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
        "/coupon/apply": {
            "post": {
                "tags": [
                    "Coupon"
                ],
                "summary": "Apply a coupon to a user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "coupon_Id": {
                                        "type": "integer"
                                    },
                                    "mobile": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "coupon applied successfully"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}